+++ include/test/tests.h
@@ -1,10 +1,12 @@
 #ifndef TESTS_H_
 #define TESTS_H_
 
+#include <stdint.h>
+
 /**
  * Runs all the tests and reports any errors.
  */
-int runTests();
+uint8_t runTests();
 
 /**
  * if the condition is false, the test fails with the given message
